requirements.txt
pymysql
Flask-SQLAlchemy
python-dotenv
flask-cors
flask
Flask-JWT
flask_restplus